#74: reworking compilation slightly
