+++ jodin-rs-vm/src/lib.rs
@@ -48,5 +48,6 @@ pub mod core_traits;
 pub mod error;
 pub mod fault;
 pub mod function_names;
+pub mod kernel;
 pub mod mvp;
 pub mod vm;
