+++ jodinc/src/compilation/c_compiler/statement_compiler.rs
@@ -1,6 +1,7 @@
 use crate::compilation::c_compiler::{CType, CValidIdentifier, Expression, Statement};
-use crate::compilation::{MicroCompiler, C99};
+use crate::compilation::C99;
 use jodin_common::ast::{JodinNode, JodinNodeType};
+use jodin_common::compilation::MicroCompiler;
 use jodin_common::core::literal::Literal;
 use jodin_common::core::tags::ResolvedIdentityTag;
 use jodin_common::core::types::TypeTag;
