+++ jodinc/src/compilation/c_compiler/top_level_declaration_compiler.rs
@@ -1,7 +1,8 @@
 use crate::compilation::c_compiler::components::TranslationUnit;
 use crate::compilation::c_compiler::{CType, CTypeCompiler, CValidIdentifier, FunctionCompiler};
-use crate::compilation::{MicroCompiler, C99};
+use crate::compilation::C99;
 use jodin_common::ast::{JodinNode, JodinNodeType};
+use jodin_common::compilation::MicroCompiler;
 use jodin_common::core::tags::ResolvedIdentityTag;
 use jodin_common::error::JodinResult;
