+++ jodinc/src/compilation/jodin_vm_compiler/asm_block.rs
@@ -1,8 +1,9 @@
 //! Contains supporting code for inserting and creating assembly code for the compiler
 
-use crate::compilation::{Compilable, Context, JodinVM, PaddedWriter};
+use crate::compilation::JodinVM;
 use crate::JodinResult;
 use itertools::Itertools;
+use jodin_common::compilation::{Compilable, Context, PaddedWriter};
 use jodin_common::identifier::Identifier;
 use jodin_common::mvp::bytecode::GetAsm;
 use jodin_common::mvp::bytecode::{Asm, Assembly, Bytecode};
