+++ jodinc/src/main.rs
@@ -1,7 +1,7 @@
 use jodin_asm::init_logging;
+use jodin_common::compilation_settings::CompilationSettings;
 use jodin_common::error::{JodinError, JodinErrorType, JodinResult};
 use jodinc::cli::JodinRsApp;
-use jodinc::compilation_settings::CompilationSettings;
 use jodinc::passes::analysis::analyze;
 use jodinc::passes::frontend::FilesToJodinNodeTool;
 use jodinc::passes::optimization::optimize;
